+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/attributes/common/ConnectionHttpAttributes.java
@@ -16,19 +16,18 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.traces.http.attributes.visitors;
+package co.elastic.apm.android.sdk.attributes.common;
 
 import co.elastic.apm.android.sdk.ElasticApmAgent;
+import co.elastic.apm.android.sdk.attributes.AttributesVisitor;
 import co.elastic.apm.android.sdk.internal.providers.LazyProvider;
 import co.elastic.apm.android.sdk.internal.services.Service;
 import co.elastic.apm.android.sdk.internal.services.network.NetworkService;
 import co.elastic.apm.android.sdk.internal.services.network.data.type.NetworkType;
-import co.elastic.apm.android.sdk.traces.http.attributes.HttpAttributesVisitor;
-import co.elastic.apm.android.sdk.traces.http.data.HttpRequest;
 import io.opentelemetry.api.common.AttributesBuilder;
 import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
 
-public class ConnectionHttpAttributes implements HttpAttributesVisitor {
+public class ConnectionHttpAttributes implements AttributesVisitor {
     private final LazyProvider<NetworkService> networkServiceProvider;
 
     public ConnectionHttpAttributes() {
@@ -36,7 +35,7 @@ public class ConnectionHttpAttributes implements HttpAttributesVisitor {
     }
 
     @Override
-    public void visit(AttributesBuilder builder, HttpRequest request) {
+    public void visit(AttributesBuilder builder) {
         NetworkType networkType = networkServiceProvider.get().getType();
         builder.put(SemanticAttributes.NET_HOST_CONNECTION_TYPE, networkType.getName());
         if (networkType.getSubTypeName() != null) {
